refactor(ethereum_api): move web specific extension
